requirements.txt
python-binance
numpy
websocket-client
pandas
matplotlib
keras
tensorflow
tqdm